-adapted newBook.html to new expanded book model. -Created new FormBook class to deliver all book info. -max number of authors restricted to 5 -added year field to book model

+++ src/main/java/misrraimsp/uned/pfg/firstmarket/service/CartServer.java
@@ -109,7 +109,7 @@ public class CartServer {
     public double getTotalPrice(Cart cart){
         double sum = 0;
         for (Item item : cart.getItems()){
-            sum += item.getQuantity() * item.getBook().getPrice();
+            sum += item.getQuantity() * item.getBook().getPrice().doubleValue();
         }
         return sum;
     }
